Move NamedElement property page to UML language package
